Remove redundant casts
This enables another batch of analyzers. The vast majority of changes were adding the rule for redundant casts. Removing them was done with the code fix for the solution.

+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/FontDialogTests.cs
@@ -715,7 +715,7 @@ public class FontDialogTests
         };
         int applyCallCount = 0;
         dialog.Apply += (sender, e) => applyCallCount++;
-        Assert.Throws<ArgumentException>(() => dialog.HookProc(IntPtr.Zero, (int)PInvoke.WM_COMMAND, (IntPtr)0x402, IntPtr.Zero));
+        Assert.Throws<ArgumentException>(() => dialog.HookProc(IntPtr.Zero, (int)PInvoke.WM_COMMAND, 0x402, IntPtr.Zero));
     }
 
     [WinFormsTheory]
